Compliance: refactor handlePartitions duplicate
